refactor(app): extract StoryElementsPanel and asset-editor callbacks (37q)
Creates useStoryElementsPanel.ts (~145 lines) to own the 5 StoryElementsPanel
callbacks: handleAddVariable, handleEditVariable, handleFindScreenDefinition,
handleHoverHighlightStart, handleHoverHighlightEnd. The hook takes blocks,
analysisResult, updateBlock/addBlock, setFileSystemTree, setHoverHighlightIds,
projectRootPath, addToast, and handleOpenEditor as explicit params.

Folds handleSaveImageMetadata, handleCopyImageToProject, handleSaveAudioMetadata,
handleCopyAudioToProject into useAssetManagement.ts (which already owns all
image/audio state). Hook receives setOpenTabs, setSecondaryOpenTabs, setActiveTabId,
setSecondaryActiveTabId as new params so it can update tab IDs on file-move.

App.tsx: removes ~290 lines of callback definitions, replaces with two
9-11 line hook calls. 3259 → 3000 lines. 686 tests pass, 0 lint warnings.
